WIP: reorganize SWR
